+++ NEWS.md
@@ -1,5 +1,7 @@
 # AzureDatalakeStore 0.0.0.9000
 
+* Adds function `adls_open_to_text()` to read file and return a character string.
+
 * Adds function `adls_rename()` to rename a file or directory.
 
 * Adds function `adls_delete()` to delete a file or directory.